+++ src/main/js/pane/boolean-input-test.ts
