+++ src/main/js/pane/boolean-monitor-test.ts
@@ -6,7 +6,7 @@ import {MultiLogMonitorController} from '../controller/monitor/multi-log';
 import {SingleLogMonitorController} from '../controller/monitor/single-log';
 import {TestUtil} from '../misc/test-util';
 import {IntervalTicker} from '../misc/ticker/interval';
-import {PlainTweakpane} from '../pane/plain-tweakpane';
+import {PlainTweakpane} from './plain-tweakpane';
 
 function createPane(): PlainTweakpane {
 	return new PlainTweakpane({
@@ -19,21 +19,21 @@ describe(PlainTweakpane.name, () => {
 		{
 			expectedClass: SingleLogMonitorController,
 			params: {},
-			value: 'foobar',
+			value: true,
 		},
 		{
 			expectedClass: MultiLogMonitorController,
 			params: {
 				count: 10,
 			},
-			value: 'foobar',
+			value: true,
 		},
 	].forEach((testCase) => {
 		context(`when params = ${JSON.stringify(testCase.params)}`, () => {
 			it(`should return class ${testCase.expectedClass.name}`, () => {
-				const api = createPane();
+				const pane = createPane();
 				const obj = {foo: testCase.value};
-				const bapi = api.addMonitor(obj, 'foo', testCase.params);
+				const bapi = pane.addMonitor(obj, 'foo', testCase.params);
 				assert.instanceOf(bapi.controller.controller, testCase.expectedClass);
 
 				const b = bapi.controller.binding;
