+++ src/main/js/pane/input-test.ts
@@ -4,7 +4,7 @@ import {describe as context, describe, it} from 'mocha';
 import {PaneError} from '../misc/pane-error';
 import {TestUtil} from '../misc/test-util';
 import {Color} from '../model/color';
-import {PlainTweakpane} from '../pane/plain-tweakpane';
+import {PlainTweakpane} from './plain-tweakpane';
 
 function createPane(): PlainTweakpane {
 	return new PlainTweakpane({
