+++ src/main/js/pane/point-2d-input-test.ts
@@ -7,7 +7,7 @@ import {StepConstraint} from '../constraint/step';
 import {ConstraintUtil} from '../constraint/util';
 import {Point2dPadTextInputController} from '../controller/input/point-2d-pad-text';
 import {TestUtil} from '../misc/test-util';
-import {PlainTweakpane} from '../pane/plain-tweakpane';
+import {PlainTweakpane} from './plain-tweakpane';
 
 function createPane(): PlainTweakpane {
 	return new PlainTweakpane({
